+++ cms/src/app/App.tsx
@@ -1,9 +1,9 @@
 import { RouterProvider } from "@tanstack/react-router";
 
 import AuthProvider, { useAuth } from "../hooks/useAuth";
-import UserProvider from "../hooks/useUser";
+import UserProvider, { useUserData } from "../hooks/useUser";
 import router from "../config/router.config";
-import React, { useCallback, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import AxiosPrivateProvider, {
   useAxiosPrivate,
   useAxiosPrivateConfig,
@@ -12,14 +12,18 @@ import AxiosPrivateProvider, {
 const InnerApp = () => {
   console.group("InnerApp running...");
   const auth = useAuth();
+  const user = useUserData();
   const axiosPrivate = useAxiosPrivate();
   console.log("auth:", auth);
+  console.log("user:", user);
   console.groupEnd();
   if (auth.isLoading) {
     return <p>Loading...</p>;
   }
 
-  return <RouterProvider router={router} context={{ auth, axiosPrivate }} />;
+  return (
+    <RouterProvider router={router} context={{ auth, axiosPrivate, user }} />
+  );
 };
 
 const AppProviders: React.FC<{ children: React.ReactNode }> = ({
@@ -27,7 +31,6 @@ const AppProviders: React.FC<{ children: React.ReactNode }> = ({
 }) => {
   console.log("[AppProviders] rendering...");
   const [accessToken, setAccessToken] = useState<string | null>(null);
-  // const getAccessToken = useCallback(() => accessToken, [accessToken]);
   const updateAccessToken = useCallback((newAccessToken: string | null) => {
     setAccessToken(newAccessToken);
   }, []);