+++ cms/src/routes/_protected/_dashboard/_user/$username/posts.tsx
@@ -1,4 +1,5 @@
 import { createFileRoute, useRouteContext } from "@tanstack/react-router";
+import Posts from "../../../../../features/posts/Posts";
 
 export const Route = createFileRoute(
   "/_protected/_dashboard/_user/$username/posts"
@@ -8,10 +9,21 @@ export const Route = createFileRoute(
     console.log("context:", context);
     console.log("params:", params);
     console.groupEnd();
-    const response = await context.axiosPrivate.get(`/users`);
-    console.log("response:", response);
+
+    if (context.auth.isAuthenticated) {
+      try {
+        // What if the username parameter is the same as the authenticated user?
+        const response = await context.axiosPrivate.get(
+          `/users/${params.username}/posts`
+        );
+        console.log(
+          `[response] for /users/${params.username}/posts:`,
+          response
+        );
+      } catch (err) {}
+    }
   },
-  component: RouteComponent,
+  component: Posts,
 });
 
 function RouteComponent() {